backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
torch>=2.6.0
transformers>=4.40.0
bitsandbytes>=0.40.0
accelerate>=0.20.0
numpy>=1.20.0
scipy>=1.10.0